passing again
Former-commit-id: 69d1f26babdad95da20bc5daee2d0d619aef3939 [formerly 19f4af1e54962e1a0d159c89be7f07520cd19476] [formerly d73d3ddf719519e65e53bbe8761a350dec546000] [formerly 140234d9f74d13890883d26ac1f6dbafbe07bf1e [formerly 219486271f71fcbe79d9960cb844347790bbcaeb] [formerly 65f0b8586a5dd9342e476fcfa71592d8ef21c068 [formerly b8a9896d61558de6cb452a2f593c97f44b973d6a]]] [formerly 114f80df6dbf654df3a22fac305290ae8f831869 [formerly 439be41715ad9350fd4487baede4d6ba606821b5] [formerly fdd83a92113dd0e7d89e626ebc6b17624c69a698] [formerly 5bb89e064a3cd509a86781edb392de36e50dd6e5 [formerly 3f430c1128e4576efeb5b3a6133688e3d9f499b9] [formerly fdaae8dc2475c19314cc66ae39624544b4dd5cf1 [formerly 24f81bc7cb041025b0d1d4aff392ac450eabc57a]]]]
Former-commit-id: 114f80df6dbf654df3a22fac305290ae8f831869 [formerly 439be41715ad9350fd4487baede4d6ba606821b5] [formerly fdd83a92113dd0e7d89e626ebc6b17624c69a698]
Former-commit-id: ad8bcaec651d1cb0324c9fee628eb8049f3f1eb3 [formerly a8a24c6aacf11b781bec329d462f04c0eca530bc] [formerly 09f71f76551b76cfa23a78340f3b4f945eff9c53 [formerly 958d7bcc8b8052b2539812d64aa28e3c0d750071]]
Former-commit-id: 4a5468cff32188abd427680110239e26f2e344c9
Former-commit-id: 980856a9f88daaa7a5b5ae6d8107610f684793f7

+++ contracts/IUSDI.sol
@@ -5,11 +5,10 @@ import "./_external/IERC20.sol";
 
 /// @title interface to interact with USDI contract
 interface IUSDI is IERC20 {
-
   function initialize(address reserveAddress) external;
 
   // getters
-  function reserveRatio() external view returns (uint256);
+  function reserveRatio() external view returns (uint192);
 
   // business
   function deposit(uint256 usdc_amount) external;
@@ -18,6 +17,7 @@ interface IUSDI is IERC20 {
 
   // admin & internal use functions
   function pause() external;
+
   function unpause() external;
 
   function mint(uint256 usdc_amount) external;
